requirements.txt
praw
mysql-connector-python
beautifulsoup4
pytesseract
Pillow
requests
nltk
scikit-learn
numpy